Move no transport code into separate header file
